requirements_onnx.txt
paddlepaddle>=2.4.0
onnxruntime>=1.10.0
opencv-python>=4.5.4
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.4.0
pyyaml>=5.0
paddle2onnx>=1.0.0
flask>=2.0.0 